# use faster gzip for backup