bzip is slow (search gzip bzip benchmark) and compresses a little bit better.

Examples:
**Larger environment**
Uncompressed: 1.16GB

time make mysql-backup  # bzip2
Creating a backup of your database in backup/mysql.sql.bz2
Backup successful!
make mysql-backup  98.77s user 1.62s system 98% cpu **1:43.25 total**
Size: 153 MB

time make mysql-backup # gzip
Creating a backup of your database in backup/mysql.sql.gz
Backup successful!
make mysql-backup  30.86s user 1.38s system 98% cpu **32.786 total**
Size: 205 MB

**Smaller environment**
Uncompressed 36MB

time make mysql-backup # gzip
Creating a backup of your database in backup/mysql.sql.gz
Backup successful!
make mysql-backup  0.94s user 0.08s system 79% cpu **1.288 total**
Size: 2.8MB

time make mysql-backup # bzip2
Creating a backup of your database in backup/mysql.sql.bz2
Backup successful!
make mysql-backup  5.08s user 0.11s system 96% cpu **5.349 total**
Size: 2.0MB

* Much faster because both use a single core.
* Restore is also faster.